Move into rules/testutils/ to match #185.
